refac(cloudfront): move AOC to sep file
